fixed ci script and moved einspline_spo.hpp
