+++ src/QMCWaveFunctions/einspline_spo.hpp
