+++ src/miniapps/einspline_spo_nested.cpp
@@ -19,7 +19,7 @@
 #include <Simulation/Simulation.hpp>
 #include <miniapps/pseudo.hpp>
 #include <Utilities/Timer.h>
-#include <miniapps/einspline_spo.hpp>
+#include <QMCWaveFunctions/einspline_spo.hpp>
 #include <getopt.h>
 
 using namespace std;
