+++ src/miniapps/miniqmc.cpp
@@ -51,7 +51,7 @@
 #include <Utilities/RandomGenerator.h>
 #include <Simulation/Simulation.hpp>
 #include <miniapps/pseudo.hpp>
-#include <miniapps/einspline_spo.hpp>
+#include <QMCWaveFunctions/einspline_spo.hpp>
 #include <QMCWaveFunctions/FakeWaveFunction.h>
 #include <getopt.h>
